fairgig-backend/core/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-jose[cryptography]
httpx
python-multipart
supabase
pydantic
python-dotenv
